+++ test/testSimplifier.ts
@@ -153,9 +153,8 @@ function tokenizeKeySequence(sequence: string): string[] {
   // no close bracket, probably trying to do a left shift, take literal
   // char sequence
   const rawTokenize = (characters: string): void => {
-    // eslint-disable-next-line @typescript-eslint/prefer-for-of
-    for (let i = 0; i < characters.length; i++) {
-      result.push(characters[i]);
+    for (const c of characters) {
+      result.push(c);
     }
   };
 
@@ -214,10 +213,7 @@ async function testIt(testObj: ITestObject): Promise<ModeHandler> {
   // Initialize the editor with the starting text and cursor selection
   assert.ok(
     await editor.edit((builder) => {
-      builder.replace(
-        new vscode.Range(new Position(0, 0), TextEditor.getDocumentEnd(editor.document)),
-        start.lines.join('\n'),
-      );
+      builder.replace(TextEditor.getDocumentRange(editor.document), start.lines.join('\n'));
     }),
     'Edit failed',
   );
@@ -253,14 +249,14 @@ async function testIt(testObj: ITestObject): Promise<ModeHandler> {
   // Check final cursor position
   const actualPosition = modeHandler.vimState.editor.selection.start;
   const expectedPosition = end.cursor;
-  assert.deepStrictEqual(
+  assert.deepEqual(
     { line: actualPosition.line, character: actualPosition.character },
     { line: expectedPosition.line, character: expectedPosition.character },
     'Cursor position is wrong.',
   );
 
   if (testObj.endMode !== undefined) {
-    assert.strictEqual(
+    assert.equal(
       Mode[modeHandler.vimState.currentMode],
       Mode[testObj.endMode],
       "Didn't enter correct mode.",
@@ -268,7 +264,7 @@ async function testIt(testObj: ITestObject): Promise<ModeHandler> {
   }
 
   if (testObj.endFsPath !== undefined) {
-    assert.strictEqual(
+    assert.equal(
       vscode.window.activeTextEditor?.document.uri.fsPath,
       typeof testObj.endFsPath === 'string' ? testObj.endFsPath : testObj.endFsPath(),
       'Active document is wrong.',
@@ -278,15 +274,15 @@ async function testIt(testObj: ITestObject): Promise<ModeHandler> {
   if (testObj.registers !== undefined) {
     for (const reg in testObj.registers) {
       if (testObj.registers[reg] !== undefined) {
-        assert.strictEqual((await Register.get(reg))?.text, testObj.registers[reg]);
+        assert.equal((await Register.get(reg))?.text, testObj.registers[reg]);
       } else {
-        assert.strictEqual(await Register.get(reg), undefined);
+        assert.equal(await Register.get(reg), undefined);
       }
     }
   }
 
   if (testObj.statusBar !== undefined) {
-    assert.strictEqual(
+    assert.equal(
       StatusBar.getText(),
       testObj.statusBar.replace('{FILENAME}', modeHandler.vimState.document.fileName),
       'Status bar text is wrong.',
@@ -296,7 +292,7 @@ async function testIt(testObj: ITestObject): Promise<ModeHandler> {
   // jumps: check jumps are correct if given
   if (testObj.jumps !== undefined) {
     // TODO: Jumps should be specified by Positions, not line contents
-    assert.deepStrictEqual(
+    assert.deepEqual(
       globalState.jumpTracker.jumps.map((j) => end.lines[j.position.line] || '<MISSING>'),
       testObj.jumps.map((t) => t.replace('|', '')),
       'Incorrect jumps found',
@@ -309,11 +305,7 @@ async function testIt(testObj: ITestObject): Promise<ModeHandler> {
       '<FRONT>';
     const expectedJumpPosition = stripBar(testObj.jumps.find((t) => t.includes('|'))) || '<FRONT>';
 
-    assert.deepStrictEqual(
-      actualJumpPosition,
-      expectedJumpPosition,
-      'Incorrect jump position found',
-    );
+    assert.deepEqual(actualJumpPosition, expectedJumpPosition, 'Incorrect jump position found');
   }
 
   return modeHandler;
@@ -454,7 +446,7 @@ async function testItWithRemaps(testObj: ITestWithRemapsObject): Promise<ModeHan
     // Lines after keys pressed but before any timeout
 
     // Check given end output is correct
-    assert.strictEqual(
+    assert.equal(
       result1.lines,
       resolvedStep.end.lines.join(os.EOL),
       `Document content does not match on step ${stepTitleOrIndex}.`,
@@ -463,7 +455,7 @@ async function testItWithRemaps(testObj: ITestWithRemapsObject): Promise<ModeHan
     // Check end cursor position
     const actualEndPosition = result1.position;
     const expectedEndPosition = resolvedStep.end.cursor;
-    assert.deepStrictEqual(
+    assert.deepEqual(
       { line: actualEndPosition.line, character: actualEndPosition.character },
       { line: expectedEndPosition.line, character: expectedEndPosition.character },
       `Cursor position is wrong on step ${stepTitleOrIndex}.`,
@@ -472,7 +464,7 @@ async function testItWithRemaps(testObj: ITestWithRemapsObject): Promise<ModeHan
     // endMode: check end mode is correct if given
     const expectedEndMode = step.stepResult.endMode;
     if (expectedEndMode !== undefined) {
-      assert.strictEqual(
+      assert.equal(
         Mode[result1.endMode],
         Mode[expectedEndMode],
         `Didn't enter correct mode on step ${stepTitleOrIndex}.`,
@@ -481,10 +473,10 @@ async function testItWithRemaps(testObj: ITestWithRemapsObject): Promise<ModeHan
 
     if (result2) {
       // After the timeout finishes (plus an offset to be sure it finished)
-      assert.notStrictEqual(result2, undefined);
+      assert.notEqual(result2, undefined);
 
       // Check given endAfterTimeout output is correct
-      assert.strictEqual(
+      assert.equal(
         result2.lines,
         resolvedStep.endAfterTimeout?.lines.join(os.EOL),
         `Document content does not match on step ${stepTitleOrIndex} after timeout.`,
@@ -493,7 +485,7 @@ async function testItWithRemaps(testObj: ITestWithRemapsObject): Promise<ModeHan
       // Check endAfterTimeout cursor position
       const actualEndAfterTimeoutPosition = result2.position;
       const expectedEndAfterTimeoutPosition = resolvedStep.endAfterTimeout!.cursor;
-      assert.deepStrictEqual(
+      assert.deepEqual(
         {
           line: actualEndAfterTimeoutPosition.line,
           character: actualEndAfterTimeoutPosition.character,
@@ -508,7 +500,7 @@ async function testItWithRemaps(testObj: ITestWithRemapsObject): Promise<ModeHan
       // endMode: check end mode is correct if given
       const expectedEndAfterTimeoutMode = step.stepResult.endModeAfterTimeout;
       if (expectedEndAfterTimeoutMode !== undefined) {
-        assert.strictEqual(
+        assert.equal(
           Mode[result2.endMode],
           Mode[expectedEndAfterTimeoutMode],
           `Didn't enter correct mode on step ${stepTitleOrIndex} after Timeout.`,